wip: do not remove if gone

+++ lib/Ravada/VM/Void.pm
@@ -425,6 +425,7 @@ sub remove_file($self, @files) {
         if $file =~ m{[`'\(\)\[]};
 
         if ($self->is_local) {
+            next if ! -e $file;
             unlink $file or die "$! $file";
         } else {
             $self->run_command("/bin/rm", $file);
